refactor: move Role and Difficulty to enums/
