+++ src/main/java/cvds/todo/backend/enums/Role.java
@@ -1,4 +1,4 @@
-package cvds.todo.backend.model;
+package cvds.todo.backend.enums;
 
 public enum Role {
     USER, ADMIN, GUESS
